+++ src/tests.zig
@@ -9,6 +9,7 @@ comptime {
     _ = @import("chess/piece.zig");
     _ = @import("chess/piece/ConditionImpl.zig");
     _ = @import("chess/piece/LookupImpl.zig");
+    _ = @import("chess/PieceList.zig");
     _ = @import("chess/Placement.zig");
     _ = @import("chess/square.zig");
 }